+++ test/clojure-mode-font-lock-test.el
@@ -942,6 +942,22 @@ DESCRIPTION is the description of the spec."
     ("\\ク"
      (1 2 clojure-character-face)))
 
+  (when-fontifying-it "should handle characters not by themselves"
+    ("[\\,,]"
+     (1 1 nil)
+     (2 3 clojure-character-face)
+     (4 5 nil))
+
+    ("[\\[]"
+     (1 1 nil)
+     (2 3 clojure-character-face)
+     (4 4 nil)))
+
+  (when-fontifying-it "should handle % character literal"
+    ("#(str \\% %)"
+     (7 8 clojure-character-face)
+     (10 10 font-lock-variable-name-face)))
+
   (when-fontifying-it "should handle referred vars"
     ("foo/var"
      (1 3 font-lock-type-face))